Avoid double rounding in experimental AAC output
Keep Glint's existing double precision through the output buffer and quantize directly to int16. A float intermediate moved a Lion Alex sample onto a rounding tie, which changed the final paragraph. Controlled single-sample replacement reproduced the discrepancy in both directions.

All 56 native configurations now match Media Foundation within two PCM units, with equal lengths and exact repeats. Add an explicit comparison gate and binary64 rounding coverage. Android audio/settings suites pass on both S22 ABIs and Watch 2; decoder defaults remain experimental.

+++ tools/aac_experiment/glint_bridge.cpp
@@ -28,7 +28,10 @@ extern "C" int pt_glint_decode(void *p, const unsigned char *data, int len,
                                 short *out, unsigned rate, unsigned channels) {
     DecoderFloatEnvironment fp;
     try {
-        float pcm[2048];
+        // The transform already uses double. Keep that precision until the
+        // one and only PCM quantization; a float intermediate can round a
+        // value onto a half-way boundary and change the final integer.
+        double pcm[2048];
         glint::aac::AacFrameInfo info;
         int n = static_cast<glint::aac::AacDecoder *>(p)->decode_frame(data, len, pcm, &info);
         if (n != 1024 || info.sample_rate != (int)rate ||
